{{cookiecutter.project_name}}/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
appnope==0.1.0            # via ipython
attrs==20.2.0             # via pytest
backcall==0.2.0           # via ipython
coverage==5.3             # via pytest-cov
decorator==4.4.2          # via ipython
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django==2.2.16
djangorestframework-jwt==1.11.0
djangorestframework==3.12.1
factory-boy==3.1.0
faker==4.14.0
gunicorn==20.0.4
importlib-metadata==2.0.0  # via pluggy, pytest
inflection==0.5.1         # via pytest-factoryboy
iniconfig==1.1.1          # via pytest
ipdb==0.13.4
ipython-genutils==0.2.0   # via traitlets
ipython==7.18.1           # via ipdb
jedi==0.17.2              # via ipython
packaging==20.4           # via pytest, pytest-sugar
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.8     # via ipython
psycopg2-binary==2.8.6
ptyprocess==0.6.0         # via pexpect
py==1.9.0                 # via pytest
pygments==2.7.2           # via ipython
pyjwt==1.7.1              # via djangorestframework-jwt
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1
pytest-django==4.1.0
pytest-factoryboy==2.0.3
pytest-instafail==0.4.2
pytest-sugar==0.9.4
pytest==6.1.2             # via pytest-cov, pytest-django, pytest-factoryboy, pytest-instafail, pytest-sugar
python-dateutil==2.8.1    # via faker
pytz==2020.1              # via django
six==1.15.0               # via packaging, python-dateutil
sqlparse==0.4.1           # via django
termcolor==1.1.0          # via pytest-sugar
text-unidecode==1.3       # via faker
toml==0.10.1              # via pytest
traitlets==5.0.5          # via ipython
wcwidth==0.2.5            # via prompt-toolkit
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
